requirements.txt
gdown==5.1.0
torch==2.2.1
pandas==2.0.3
pydantic==2.7.1
accelerate==0.30.0
peft==0.10.0
transformers
yacs
ogb
numpy
python-graphblas
datasets
accelerate
pytest